(PA-7665) Add logic to run in puppet-agent repo
Prior to this commit, Ticketmatch ran a Git clone on the puppet-agent
repository at the specified revision.

This commit adds a conditional to check whether the script is running in
the puppet-agent repository already and skip the initial clone if it is.
This is helpful for running Ticketmatch in the puppet-agent repository
as a GitHub Action.